+++ backend/src/auth/auth.module.ts
@@ -1,0 +1,26 @@
+import { Module } from '@nestjs/common';
+import { AuthService } from './auth.service';
+import { AuthController } from './auth.controller';
+import { TypeOrmExModule } from '../typeorm/typeorm-ex.module';
+import { UsersRepository } from '../users/users.repository';
+import { JwtModule } from '@nestjs/jwt';
+import { PassportModule } from '@nestjs/passport';
+import { JwtStrategy } from './jwt.strategy';
+
+@Module({
+  controllers: [AuthController],
+  providers: [AuthService, JwtStrategy],
+  imports: [
+    PassportModule.register({ defaultStrategy: 'jwt' }),
+    JwtModule.register({
+      secret: 'Secret1234',
+      signOptions: {
+        expiresIn: 3600,
+      },
+    }),
+    TypeOrmExModule.forCustomRepository([UsersRepository]), // CustomRepository 를 사용하기 위해 추가
+  ],
+  // 다른곳에서도 jwt 인증을 사용하기 위해 export
+  exports: [JwtStrategy, PassportModule],
+})
+export class AuthModule {}
